Assignment 1
In this assignment you will develop four HTML pages, according to the following specifications.

Choose a content sharing domain of interest to you. You must develop your own unique site with branding, categories, and items representative of that domain. Creativity is encouraged.  

You are not allowed to select previously created branding / content. You must create your own. In particular, you must not use any branding / content from (1) the assignment itself; (2) the textbook; or (3) any other students in any of the sections of the course.  
Each of the pages specified must be a standalone, individual HTML page. And not, for example, framed or one single page with main content section visibility switching.  
Each of the pages must be standalone, but must have a common design to preserve the look-and-feel of the site. So, some of the common look-and-feel elements will be repeated across pages. Common look-and-feel elements must render exactly the same, so there are seamless transitions with no jitter between pages for common elements.  
You must conform to the required overall page layout structure.  
Within each of the sections of the overall page layout structure, you may style the content differently than the representative examples (color scheme, alignment, etc.), as long as the illustrated types of content are present and represented in a reasonable and understandable way.  
All of the content must be represented using HTML directly (no JavaScript DOM generation).  
For this assignment, we will consider the following specific design standard requirements:
Validation of HTML5 by the following tool: http://html5.validator.nu/  
Cross-browser checking for Firefox and Chrome  